pyproject.toml
[tool.poetry]
name = "pollsome"
version = "0.1.0"
description = "An elegant Polling discord bot made with discord.py"
authors = ["Yash Prakash <yashprakash13@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.0.1"
black = "^22.8.0"
isort = "^5.10.1"
python-dotenv = "^0.21.0"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
